requirements.txt
Alfred-Workflow==1.36
xkcdpass==1.16.5
